updat

+++ my_merge.py
@@ -8,6 +8,7 @@ from transformers import AutoModelForCausalLM, AutoTokenizer
 from peft import PeftModel 
 from model_merging_methods.merging_methods import MergingMethod
 from utils.utils import set_random_seed, smart_tokenizer_and_embedding_resize
+import torch
 
 
 def main():
@@ -23,13 +24,13 @@ def main():
 
     print(args)
 
-    base_tokenizer = AutoTokenizer.from_pretrained(args.llama_path, use_fast=False)
+    base_tokenizer = AutoTokenizer.from_pretrained(args.llama_path, use_fast=False, dtype=torch.float16)
     base_model = AutoModelForCausalLM.from_pretrained(args.llama_path, cache_dir="../models/")
 
     tokenizer1 = AutoTokenizer.from_pretrained(args.m1, use_fast=False)
     tokenizer2 = AutoTokenizer.from_pretrained(args.m2, use_fast=False)
-    model1 = AutoModelForCausalLM.from_pretrained(args.m1)
-    model2 = AutoModelForCausalLM.from_pretrained(args.m2)
+    model1 = AutoModelForCausalLM.from_pretrained(args.m1, dtype=torch.float16)
+    model2 = AutoModelForCausalLM.from_pretrained(args.m2, dtype=torch.float16)
 
     smart_tokenizer_and_embedding_resize(
         special_tokens_dict=dict(pad_token="[PAD]"),
